+++ src/main/java/com/aleksander/restaurant/reservation/controller/TableController.java
@@ -9,6 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.aleksander.restaurant.reservation.dto.TableDTO;
@@ -28,10 +29,11 @@ public class TableController {
 
     @GetMapping
     public ResponseEntity<List<TableDTO>> getTables(
-            @ParameterObject TableFilter filter) {
+            @ParameterObject TableFilter filter,
+            @RequestParam(defaultValue = "false") boolean recommend) {
 
         return ResponseEntity.ok(
-                tableService.findTables(filter));
+                tableService.findTables(filter, recommend));
     }
 
     @GetMapping("/{tableNumber}/availability")